feat:Initialize Koin in Application class
This commit introduces the `BookApplication` class, which extends `android.app.Application`.
The primary purpose of this class is to initialize Koin dependency injection by calling `initKoin` in its `onCreate` method.
The `androidContext` is provided to Koin.

Additionally, the `AndroidManifest.xml` has been updated to specify `.BookApplication` as the `android:name` for the `<application>` tag.