feat:Create PulseAnimation composable
This commit introduces a new composable function `PulseAnimation`.

This composable creates a visual pulsing effect. It uses `rememberInfiniteTransition` and `animateFloat` to animate the scale and alpha of a `Box`. The `Box` has a circular border with the color `SandYellow`.

The animation repeats infinitely, scaling from 0 to 1 and fading out as it scales up.